+++ forui/lib/src/theme/theme_data.dart
@@ -293,6 +293,17 @@ final class FThemeData with Diagnosticable, _$FThemeDataFunctions {
   @override
   final FMultiSelectStyle multiSelectStyle;
 
+  /// The modal sheet style.
+  ///
+  /// ## CLI
+  /// To generate and customize this style:
+  ///
+  /// ```shell
+  /// dart run forui style create modal-sheet
+  /// ```
+  @override
+  final FModalSheetStyle modalSheetStyle;
+
   /// The pagination style.
   ///
   /// ## CLI
@@ -303,6 +314,17 @@ final class FThemeData with Diagnosticable, _$FThemeDataFunctions {
   /// ```
   @override
   final FPaginationStyle paginationStyle;
+
+  /// The persistent sheet style.
+  ///
+  /// ## CLI
+  /// To generate and customize this style:
+  ///
+  /// ```shell
+  /// dart run forui style create persistent-sheet
+  /// ```
+  @override
+  final FPersistentSheetStyle persistentSheetStyle;
 
   /// The picker's style.
   ///
@@ -413,17 +435,6 @@ final class FThemeData with Diagnosticable, _$FThemeDataFunctions {
   /// ```
   @override
   final FSelectMenuTileStyle selectMenuTileStyle;
-
-  /// The sheet style.
-  ///
-  /// ## CLI
-  /// To generate and customize this style:
-  ///
-  /// ```shell
-  /// dart run forui style create sheet
-  /// ```
-  @override
-  final FSheetStyle sheetStyle;
 
   /// The sidebar style.
   ///
@@ -589,7 +600,9 @@ final class FThemeData with Diagnosticable, _$FThemeDataFunctions {
     FLabelStyles? labelStyles,
     FLineCalendarStyle? lineCalendarStyle,
     FMultiSelectStyle? multiSelectStyle,
+    FModalSheetStyle? modalSheetStyle,
     FPaginationStyle? paginationStyle,
+    FPersistentSheetStyle? persistentSheetStyle,
     FPickerStyle? pickerStyle,
     FPopoverStyle? popoverStyle,
     FPopoverMenuStyle? popoverMenuStyle,
@@ -600,7 +613,6 @@ final class FThemeData with Diagnosticable, _$FThemeDataFunctions {
     FSelectStyle? selectStyle,
     FSelectGroupStyle? selectGroupStyle,
     FSelectMenuTileStyle? selectMenuTileStyle,
-    FSheetStyle? sheetStyle,
     FSidebarStyle? sidebarStyle,
     FSliderStyles? sliderStyles,
     FToasterStyle? toasterStyle,
@@ -653,8 +665,10 @@ final class FThemeData with Diagnosticable, _$FThemeDataFunctions {
           lineCalendarStyle ?? FLineCalendarStyle.inherit(colors: colors, typography: typography, style: style),
       multiSelectStyle:
           multiSelectStyle ?? FMultiSelectStyle.inherit(colors: colors, typography: typography, style: style),
+      modalSheetStyle: modalSheetStyle ?? FModalSheetStyle.inherit(colors: colors),
       paginationStyle:
           paginationStyle ?? FPaginationStyle.inherit(colors: colors, typography: typography, style: style),
+      persistentSheetStyle: persistentSheetStyle ?? const FPersistentSheetStyle(),
       pickerStyle: pickerStyle ?? FPickerStyle.inherit(colors: colors, style: style, typography: typography),
       popoverStyle: popoverStyle ?? FPopoverStyle.inherit(colors: colors, style: style),
       popoverMenuStyle:
@@ -668,7 +682,6 @@ final class FThemeData with Diagnosticable, _$FThemeDataFunctions {
           selectGroupStyle ?? FSelectGroupStyle.inherit(colors: colors, typography: typography, style: style),
       selectMenuTileStyle:
           selectMenuTileStyle ?? FSelectMenuTileStyle.inherit(colors: colors, typography: typography, style: style),
-      sheetStyle: sheetStyle ?? FSheetStyle.inherit(colors: colors),
       sidebarStyle: sidebarStyle ?? FSidebarStyle.inherit(colors: colors, typography: typography, style: style),
       sliderStyles: sliderStyles ?? FSliderStyles.inherit(colors: colors, typography: typography, style: style),
       toasterStyle: toasterStyle ?? FToasterStyle.inherit(colors: colors, typography: typography, style: style),
@@ -716,7 +729,9 @@ final class FThemeData with Diagnosticable, _$FThemeDataFunctions {
     labelStyles: a.labelStyles.lerp(b.labelStyles, t),
     lineCalendarStyle: a.lineCalendarStyle.lerp(b.lineCalendarStyle, t),
     multiSelectStyle: a.multiSelectStyle.lerp(b.multiSelectStyle, t),
+    modalSheetStyle: a.modalSheetStyle.lerp(b.modalSheetStyle, t),
     paginationStyle: a.paginationStyle.lerp(b.paginationStyle, t),
+    persistentSheetStyle: a.persistentSheetStyle.lerp(b.persistentSheetStyle, t),
     pickerStyle: a.pickerStyle.lerp(b.pickerStyle, t),
     popoverStyle: a.popoverStyle.lerp(b.popoverStyle, t),
     popoverMenuStyle: a.popoverMenuStyle.lerp(b.popoverMenuStyle, t),
@@ -727,7 +742,6 @@ final class FThemeData with Diagnosticable, _$FThemeDataFunctions {
     selectStyle: a.selectStyle.lerp(b.selectStyle, t),
     selectGroupStyle: a.selectGroupStyle.lerp(b.selectGroupStyle, t),
     selectMenuTileStyle: a.selectMenuTileStyle.lerp(b.selectMenuTileStyle, t),
-    sheetStyle: a.sheetStyle.lerp(b.sheetStyle, t),
     sidebarStyle: a.sidebarStyle.lerp(b.sidebarStyle, t),
     sliderStyles: a.sliderStyles.lerp(b.sliderStyles, t),
     toasterStyle: a.toasterStyle.lerp(b.toasterStyle, t),
@@ -775,7 +789,9 @@ final class FThemeData with Diagnosticable, _$FThemeDataFunctions {
     required this.labelStyles,
     required this.lineCalendarStyle,
     required this.multiSelectStyle,
+    required this.modalSheetStyle,
     required this.paginationStyle,
+    required this.persistentSheetStyle,
     required this.pickerStyle,
     required this.popoverStyle,
     required this.popoverMenuStyle,
@@ -786,7 +802,6 @@ final class FThemeData with Diagnosticable, _$FThemeDataFunctions {
     required this.selectStyle,
     required this.selectGroupStyle,
     required this.selectMenuTileStyle,
-    required this.sheetStyle,
     required this.sidebarStyle,
     required this.sliderStyles,
     required this.toasterStyle,
@@ -1274,7 +1289,9 @@ final class FThemeData with Diagnosticable, _$FThemeDataFunctions {
     FLabelStyles Function(FLabelStyles style)? labelStyles,
     FLineCalendarStyle Function(FLineCalendarStyle style)? lineCalendarStyle,
     FMultiSelectStyle Function(FMultiSelectStyle style)? multiSelectStyle,
+    FModalSheetStyle Function(FModalSheetStyle style)? modalSheetStyle,
     FPaginationStyle Function(FPaginationStyle style)? paginationStyle,
+    FPersistentSheetStyle Function(FPersistentSheetStyle style)? persistentSheetStyle,
     FPickerStyle Function(FPickerStyle style)? pickerStyle,
     FPopoverStyle Function(FPopoverStyle style)? popoverStyle,
     FPopoverMenuStyle Function(FPopoverMenuStyle style)? popoverMenuStyle,
@@ -1285,7 +1302,6 @@ final class FThemeData with Diagnosticable, _$FThemeDataFunctions {
     FSelectStyle Function(FSelectStyle style)? selectStyle,
     FSelectGroupStyle Function(FSelectGroupStyle style)? selectGroupStyle,
     FSelectMenuTileStyle Function(FSelectMenuTileStyle style)? selectMenuTileStyle,
-    FSheetStyle Function(FSheetStyle style)? sheetStyle,
     FSidebarStyle Function(FSidebarStyle style)? sidebarStyle,
     FSliderStyles Function(FSliderStyles style)? sliderStyles,
     FToasterStyle Function(FToasterStyle style)? toasterStyle,
@@ -1334,7 +1350,11 @@ final class FThemeData with Diagnosticable, _$FThemeDataFunctions {
     labelStyles: labelStyles != null ? labelStyles(this.labelStyles) : this.labelStyles,
     lineCalendarStyle: lineCalendarStyle != null ? lineCalendarStyle(this.lineCalendarStyle) : this.lineCalendarStyle,
     multiSelectStyle: multiSelectStyle != null ? multiSelectStyle(this.multiSelectStyle) : this.multiSelectStyle,
+    modalSheetStyle: modalSheetStyle != null ? modalSheetStyle(this.modalSheetStyle) : this.modalSheetStyle,
     paginationStyle: paginationStyle != null ? paginationStyle(this.paginationStyle) : this.paginationStyle,
+    persistentSheetStyle: persistentSheetStyle != null
+        ? persistentSheetStyle(this.persistentSheetStyle)
+        : this.persistentSheetStyle,
     pickerStyle: pickerStyle != null ? pickerStyle(this.pickerStyle) : this.pickerStyle,
     popoverStyle: popoverStyle != null ? popoverStyle(this.popoverStyle) : this.popoverStyle,
     popoverMenuStyle: popoverMenuStyle != null ? popoverMenuStyle(this.popoverMenuStyle) : this.popoverMenuStyle,
@@ -1347,7 +1367,6 @@ final class FThemeData with Diagnosticable, _$FThemeDataFunctions {
     selectMenuTileStyle: selectMenuTileStyle != null
         ? selectMenuTileStyle(this.selectMenuTileStyle)
         : this.selectMenuTileStyle,
-    sheetStyle: sheetStyle != null ? sheetStyle(this.sheetStyle) : this.sheetStyle,
     sidebarStyle: sidebarStyle != null ? sidebarStyle(this.sidebarStyle) : this.sidebarStyle,
     sliderStyles: sliderStyles != null ? sliderStyles(this.sliderStyles) : this.sliderStyles,
     toasterStyle: toasterStyle != null ? toasterStyle(this.toasterStyle) : this.toasterStyle,