+++ samples/lib/widgets/modal_sheet.dart
@@ -64,7 +64,7 @@ class BlurredModalSheetPage extends Sample {
     child: FButton(
       child: const Text('Open'),
       onPress: () => showFSheet(
-        style: context.theme.sheetStyle.copyWith(
+        style: context.theme.modalSheetStyle.copyWith(
           barrierFilter: (animation) => ImageFilter.compose(
             outer: ImageFilter.blur(sigmaX: animation * 5, sigmaY: animation * 5),
             inner: ColorFilter.mode(context.theme.colors.barrier, BlendMode.srcOver),